requirements.txt
clang
tiktoken
pandas>=2.0.0
lxml>=4.9.0
gitpython>=3.1.0
python-dotenv>=1.0.0
tabulate
pyyaml
pydot
networkx
matplotlib
intervaltree
cxxfilt
nltk
seaborn
modal
scipy
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
langchain_community
numexpr
langsmith
